default arguments
I wanted a way for people to customize the default /rand load command and decided it would be nice to default any command argument.  This has many use cases but at it's most basic, it will let people run any aribtrary set of commands on startup which can then be accessed with the up/down arrow keys.