ping clients who haven't finished logging in
I noticed that we were only pinging clients that have successfully logged all the way in.  Maybe it will help to ping clients that are having troubles connecting?  Or maybe it won't help at all.  We'll find out.

I made it so that it only sends a 1-way "poke" to those clients like we do with in-game players.