more fixes to auto-LT
the biggest problem was that it would use the local player's latencyTolerance
setting as the max value instead of what the server sent over.  This
could result in weird stuff happening when the LT started changing.
Also, a value of 0 or less is ignored in case somebody unknowingly
tries to force it to go faster than it can.