infinite LT bug
A rare bug, if you start up with the -s option then latencyTolerance *could* end up in a bad state where it would display as "inf" because it was not being initialized properly.

I also took the opportunity to make the setting of LT more robust.

Bonus fix: format LT with the proper number of digits for the frame rate.