more LT tweaks
The main change is to not increase LT when players are getting smooth results and not having to wait for packets very often.  Sometimes RTT math would suggest an increase but reality says otherwise.

Also, trying to get that starting LT value more accurate by playing with RTT calculation on pings before the game starts.