pass RTT on start/resume and code cleanup
When the server is not playing we shouldn't use their idea of LT (for autoLT).  So the kpStartRequest now contains an initial LT estimate from a player in the game and is no longer passed as part of the server Config.

Also did a fairly big code refactor to make the start/resume logic easier to understand.  Added kpStartRequest & kpResumeRequest commands which are sent by a client to the server.  Then the server calls kpStartLevel or kpResumeLevel to tell everyone that the game is starting.  This makes the code flow much easier to understand and debug.