dynamic LT
This simple-looking fix allows the effective LT to adjust with each frame based on whether the client had to wait for packets.  It will then pre-send up to 2 packets depending on how far behind the clock gets.

While this started as a fix to help out with LT=0, it should also help out with "marginal" LT settings to push those up to the proper level in real time.