Fix doors, make Teleporters always dynamic
Teleporters appear to have a bug surrounding the kShowAlwaysOption flag,
and a bug that I'm extremely reluctant to fix for backwards
compatibility reasons. The flag is actually set by default, but the game
does not actually honor it--or at least it doesn't do what it's meant to
do. "Stopped" teleporters disappear, making the flag completely useless.