sqlite3 for recent levels
This is our initial foray into using SQLite to save data.  The "recent levels" data is a fairly benign in that it's not going to be a big deal if the data were to get corrupted.  Which makes it a relatively safe area to start using a "new" technology to see how it goes.

The top-level interface to access the data store is currently named CRUD and the SQLite implementation of that is in CRUDsqlite.  This name (CRUD) may change.

As part of this change all of the level-related meta data was moved to a small LevelInfo class that is passed around instead of all the various level variables.  That class maps to the "levels" table that is created by this change.