sqlite3_busy_timeout 5msec
When testing with multiple clients at once (up to 6), I found that sometimes clients would not get the data they need from the database.  Adding a max 5msec busy timeout seems to solve that problem.  This really isn't a problem for a single client but this change shouldn't affect that scenario (there's nobody else to wait for).