Preserve behavior of wallYon = defaultYon
This was common practice back in the day to reset yon after altering it.
In truth, because it wasn't an actual variable, it was actually setting
it to 0 and therefore not setting private yon!

Without *these* changes, that would still effectively work as before,
but it would break the recent static geometry code and render everything
afterwards as dynamic.