FRandSeed random when host not playing
What looked like a problem with HUD kills sticking around too long was actually due to the fact that the currentGameId (which is set from FRandSeed) was not changing when the host was out.  This makes that random by setting a new random value on game load, regardless if the host is playing or not.