actually fixed the connection issue?
I discovered that really you can't rely on the incoming host OR port matching what you connect out on.  This fix forces the first sent message to append it's connection ID to the packet so that it can be matched with a connection that hasn't received any packets.  Then the connection address is modified to match that first packet.