Fix punch handling (#500)
* fix crash server to client switch bug

There was a bug when switching from serving to a client that was caused by the Punch callback logic.  I refactored so that the callback is using a valid "this" pointer.  Also refactored/documented punch stuff a little.

* actually fixed the connection issue?

I discovered that really you can't rely on the incoming host OR port matching what you connect out on.  This fix forces the first sent message to append it's connection ID to the packet so that it can be matched with a connection that hasn't received any packets.  Then the connection address is modified to match that first packet.

* connection table code refactor and fix

made the code easier to ready (hopefully), and fixed the check for matching a new incoming connection with the correct UDPConnection that hasn't received any packets.

* allow the connection ip/port to change early

I found that an initial UDP connection can change where it's coming from early on.  Even locally I saw some packets come from port 1024 on the router before switching to the remote host/port combo (presumably because of latent effects from punch?).

This modification allows the connection info to change for the first 100 serial numbers (50 messages) by which time it seems to be pretty stable.

* verify connection if sender provided

If we get sender in the packet then use it to verify the connection because it's possible some routers may use the same port for multiple incoming packets before the Punch magic takes hold.

* subtle bug fix

there was a condition, if packet didn't match but came from a server, that it wouldn't attempt to be processed and not released.
If we're **really** lucky, this might get servers behind double NAT to work?

* Jab->HolePunch and pass connection ID

Passing the connection ID on the HolePunch allows the punchee to match up the message with the correct connection object.

* -p option on multiplayer.sh script to specify starting client ports

Handy if port forwarding enabled in router.