Profiling tests to see why they're slow (#502)
* Remove *more* app->Get calls from main loop

Profiling the tests in Xcode revealed a bottleneck in app->Get(App->Get(kHUDPreset).  I followed Dan's lead and stored this setting in the prefs callback.

* use json references instead of copies

This optimization in `IBSPPart()` make a big difference in test speed because the construction/destruction of all the BSPs was significant with the `nlohmann::json` copies.

* don't attempt to open the audio device for tests